added mongodb connection info to config var, switched to in-memory parsing, switched to .env file for config loading, added procfile for foreman/heroku support, modified matchprovider slightly to auto-detect directory
