+++ .gitignore
@@ -1,3 +1,8 @@
 config.js
+.env
 node_modules
 replays
+*.log
+servers
+*.dem
+*.bz2
