--- config_example.js
@@ -1,23 +1,0 @@
-var config = {};
-
-config.steam_name = "";
-config.steam_user = "";
-config.steam_pass = "";
-config.steam_guard_code = "";
-config.cwd = "";
-config.steam_api_key = "";
-
-config.account_ids = [""];
-
-config.matchCount = 5;
-
-config.replaysFolder = "";
-
-config.parserPath = "";
-
-config.logFile = "";
-
-config.request_timeout = 1000 * 30;
-config.steam_response_timeout = 1000 * 30;
-
-module.exports = config;
