+++ servers
@@ -1,1 +1,1 @@
-[{"host":"208.78.164.13","port":27017},{"host":"208.78.164.14","port":27018},{"host":"208.78.164.12","port":27019},{"host":"208.78.164.10","port":27017},{"host":"208.78.164.13","port":27019},{"host":"208.78.164.14","port":27019},{"host":"208.78.164.9","port":27019},{"host":"208.78.164.10","port":27019},{"host":"208.78.164.11","port":27018},{"host":"208.78.164.11","port":27019},{"host":"208.78.164.12","port":27018},{"host":"208.78.164.13","port":27018},{"host":"208.78.164.9","port":27018},{"host":"208.78.164.10","port":27018},{"host":"208.78.164.11","port":27017},{"host":"208.78.164.12","port":27017},{"host":"208.78.164.14","port":27017},{"host":"208.78.164.9","port":27017},{"host":"208.64.200.201","port":27020},{"host":"72.165.61.185","port":27017},{"host":"208.64.201.169","port":27019},{"host":"208.64.201.169","port":27020},{"host":"208.64.200.204","port":27017},{"host":"72.165.61.175","port":27017},{"host":"208.64.200.137","port":27019},{"host":"208.64.200.137","port":27018},{"host":"208.64.200.202","port":27018},{"host":"208.64.201.176","port":27020},{"host":"208.64.200.203","port":27018},{"host":"208.64.201.176","port":27018},{"host":"72.165.61.174","port":27018},{"host":"208.64.200.204","port":27019},{"host":"72.165.61.175","port":27018},{"host":"208.64.200.205","port":27018},{"host":"208.64.200.201","port":27018},{"host":"72.165.61.174","port":27017},{"host":"208.64.200.137","port":27017},{"host":"208.64.201.176","port":27021},{"host":"208.64.201.169","port":27021},{"host":"208.64.201.176","port":27019},{"host":"208.64.200.202","port":27017},{"host":"208.64.200.201","port":27017},{"host":"208.64.200.205","port":27017},{"host":"208.64.200.203","port":27017},{"host":"208.64.200.202","port":27019},{"host":"208.64.201.169","port":27017},{"host":"208.64.200.203","port":27019},{"host":"72.165.61.186","port":27017},{"host":"208.64.201.169","port":27018},{"host":"72.165.61.187","port":27017},{"host":"208.64.201.176","port":27017},{"host":"208.64.200.204","port":27018},{"host":"208.64.200.201","port":27019},{"host":"72.165.61.188","port":27017},{"host":"208.64.200.137","port":27020},{"host":"208.64.200.205","port":27019},{"host":"81.171.115.34","port":27019},{"host":"81.171.115.37","port":27019},{"host":"81.171.115.36","port":27017},{"host":"81.171.115.35","port":27017},{"host":"81.171.115.37","port":27018},{"host":"81.171.115.35","port":27019},{"host":"81.171.115.35","port":27018},{"host":"81.171.115.36","port":27019},{"host":"81.171.115.36","port":27018},{"host":"81.171.115.34","port":27018},{"host":"81.171.115.34","port":27017},{"host":"81.171.115.37","port":27017},{"host":"146.66.152.14","port":27019},{"host":"146.66.152.15","port":27018},{"host":"146.66.152.12","port":27019},{"host":"146.66.152.15","port":27019},{"host":"146.66.152.11","port":27018},{"host":"146.66.152.11","port":27017},{"host":"146.66.152.14","port":27017},{"host":"146.66.152.10","port":27017},{"host":"146.66.152.13","port":27017},{"host":"146.66.152.12","port":27017},{"host":"146.66.152.11","port":27019},{"host":"146.66.152.15","port":27017}]
+[{"host":"208.78.164.9","port":27019},{"host":"208.78.164.9","port":27017},{"host":"208.78.164.13","port":27018},{"host":"208.78.164.10","port":27017},{"host":"208.78.164.12","port":27018},{"host":"208.78.164.11","port":27017},{"host":"208.78.164.14","port":27017},{"host":"208.78.164.13","port":27019},{"host":"208.78.164.14","port":27018},{"host":"208.78.164.10","port":27018},{"host":"208.78.164.9","port":27018},{"host":"208.78.164.12","port":27017},{"host":"208.78.164.11","port":27018},{"host":"208.78.164.11","port":27019},{"host":"208.78.164.10","port":27019},{"host":"208.78.164.13","port":27017},{"host":"208.78.164.14","port":27019},{"host":"208.78.164.12","port":27019},{"host":"208.64.201.169","port":27021},{"host":"208.64.200.201","port":27018},{"host":"208.64.200.201","port":27017},{"host":"72.165.61.174","port":27017},{"host":"208.64.201.176","port":27019},{"host":"208.64.200.202","port":27017},{"host":"208.64.200.204","port":27017},{"host":"208.64.201.169","port":27019},{"host":"208.64.200.137","port":27018},{"host":"208.64.200.204","port":27019},{"host":"208.64.201.176","port":27017},{"host":"208.64.200.202","port":27018},{"host":"208.64.200.205","port":27019},{"host":"208.64.200.203","port":27018},{"host":"208.64.201.176","port":27021},{"host":"208.64.200.204","port":27018},{"host":"208.64.201.169","port":27018},{"host":"72.165.61.185","port":27017},{"host":"72.165.61.174","port":27018},{"host":"208.64.200.137","port":27019},{"host":"208.64.201.176","port":27018},{"host":"72.165.61.175","port":27018},{"host":"72.165.61.187","port":27017},{"host":"208.64.200.137","port":27020},{"host":"72.165.61.175","port":27017},{"host":"208.64.200.201","port":27019},{"host":"208.64.200.203","port":27019},{"host":"208.64.200.203","port":27017},{"host":"208.64.200.205","port":27017},{"host":"208.64.201.169","port":27020},{"host":"208.64.201.176","port":27020},{"host":"208.64.200.205","port":27018},{"host":"208.64.200.137","port":27017},{"host":"208.64.200.201","port":27020},{"host":"72.165.61.186","port":27017},{"host":"208.64.201.169","port":27017},{"host":"72.165.61.188","port":27017},{"host":"208.64.200.202","port":27019},{"host":"81.171.115.36","port":27018},{"host":"81.171.115.36","port":27019},{"host":"81.171.115.34","port":27018},{"host":"81.171.115.35","port":27019},{"host":"81.171.115.36","port":27017},{"host":"81.171.115.34","port":27017},{"host":"81.171.115.37","port":27019},{"host":"81.171.115.37","port":27018},{"host":"81.171.115.34","port":27019},{"host":"81.171.115.35","port":27017},{"host":"81.171.115.35","port":27018},{"host":"81.171.115.37","port":27017},{"host":"146.66.152.14","port":27018},{"host":"146.66.152.13","port":27018},{"host":"146.66.152.12","port":27019},{"host":"146.66.152.10","port":27019},{"host":"146.66.152.13","port":27017},{"host":"146.66.152.11","port":27018},{"host":"146.66.152.10","port":27018},{"host":"146.66.152.15","port":27019},{"host":"146.66.152.13","port":27019},{"host":"146.66.152.14","port":27017},{"host":"146.66.152.15","port":27018},{"host":"146.66.152.11","port":27019}]
